+++ lib/take_picture.dart
@@ -1,0 +1,115 @@
+import 'package:camera/camera.dart';
+import 'package:flutter/material.dart';
+
+import 'display_picture.dart';
+
+/// 写真撮影画面
+class TakePictureScreen extends StatefulWidget {
+  const TakePictureScreen({
+    Key? key,
+    required this.camera,
+  }) : super(key: key);
+
+  final CameraDescription camera;
+
+  @override
+  TakePictureScreenState createState() => TakePictureScreenState();
+}
+
+class TakePictureScreenState extends State<TakePictureScreen> {
+  late CameraController _controller;
+  late Future<void> _initializeControllerFuture;
+  List<String> imageList = [];
+
+  @override
+  void initState() {
+    super.initState();
+
+    _controller = CameraController(
+      // カメラを指定
+      widget.camera,
+      // 解像度を定義
+      ResolutionPreset.medium,
+    );
+
+    // コントローラーを初期化
+    _initializeControllerFuture = _controller.initialize();
+  }
+
+  @override
+  void dispose() {
+    // ウィジェットが破棄されたら、コントローラーを破棄
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: FutureBuilder<void>(
+          future: _initializeControllerFuture,
+          builder: (context, snapshot) {
+            if (snapshot.connectionState == ConnectionState.done) {
+              return CameraPreview(_controller);
+            } else {
+              return const CircularProgressIndicator();
+            }
+          },
+        ),
+      ),
+      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
+      floatingActionButton: FloatingActionButton(
+        onPressed: () async {
+          // 写真を撮る
+          final image = await _controller.takePicture();
+          // 写真をリストに格納
+          imageList.add(image.path);
+        },
+        child: const Icon(Icons.camera_alt),
+      ),
+      bottomNavigationBar: BottomAppBar(
+        shape: const CircularNotchedRectangle(),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.end,
+          children: <Widget>[
+            Padding(
+              padding: const EdgeInsets.all(16.0),
+              child: IconButton(
+                icon: const Icon(Icons.photo),
+                onPressed: () {
+                  //処理
+                  if (imageList.isNotEmpty){
+                    Navigator.of(context).push(
+                      MaterialPageRoute(
+                        builder: (context) => DisplayPictureScreen(imagePathList: imageList),
+                        fullscreenDialog: true,
+                      ),
+                    );
+                  } else {
+                    showDialog(
+                      context: context,
+                      builder: (BuildContext context) {
+                        return AlertDialog(
+                          title: const Text('No Picture'),
+                          actions: <Widget>[
+                            TextButton(
+                              child: const Text('OK'),
+                              onPressed: () {
+                                Navigator.of(context).pop();
+                              },
+                            ),
+                          ],
+                        );
+                      },
+                    );
+                  }
+                },
+              ),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
